+++ src/uucore/src/lib/features/mode.rs
@@ -7,6 +7,8 @@
 
 // spell-checker:ignore (vars) fperm srwx
 
+use libc::{mode_t, S_IRGRP, S_IROTH, S_IRUSR, S_IWGRP, S_IWOTH, S_IWUSR};
+
 pub fn parse_numeric(fperm: u32, mut mode: &str) -> Result<u32, String> {
     let (op, pos) = parse_op(mode, Some('='))?;
     mode = mode[pos..].trim().trim_start_matches('0');
@@ -128,4 +130,42 @@ fn parse_change(mode: &str, fperm: u32, considering_dir: bool) -> (u32, usize) {
         srwx = 0;
     }
     (srwx, pos)
+}
+
+pub fn parse_mode(mode: Option<String>) -> Result<mode_t, String> {
+    let fperm = S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP | S_IROTH | S_IWOTH;
+    if let Some(mode) = mode {
+        let arr: &[char] = &['0', '1', '2', '3', '4', '5', '6', '7', '8', '9'];
+        let result = if mode.contains(arr) {
+            parse_numeric(fperm as u32, mode.as_str())
+        } else {
+            parse_symbolic(fperm as u32, mode.as_str(), true)
+        };
+        result.map(|mode| mode as mode_t)
+    } else {
+        Ok(fperm)
+    }
+}
+
+#[cfg(test)]
+mod test {
+
+    #[test]
+    fn symbolic_modes() {
+        assert_eq!(super::parse_mode(Some("u+x".to_owned())).unwrap(), 0o766);
+        assert_eq!(
+            super::parse_mode(Some("+x".to_owned())).unwrap(),
+            if !crate::os::is_wsl_1() { 0o777 } else { 0o776 }
+        );
+        assert_eq!(super::parse_mode(Some("a-w".to_owned())).unwrap(), 0o444);
+        assert_eq!(super::parse_mode(Some("g-r".to_owned())).unwrap(), 0o626);
+    }
+
+    #[test]
+    fn numeric_modes() {
+        assert_eq!(super::parse_mode(Some("644".to_owned())).unwrap(), 0o644);
+        assert_eq!(super::parse_mode(Some("+100".to_owned())).unwrap(), 0o766);
+        assert_eq!(super::parse_mode(Some("-4".to_owned())).unwrap(), 0o662);
+        assert_eq!(super::parse_mode(None).unwrap(), 0o666);
+    }
 }
