+++ src/uucore/src/lib/mods.rs
@@ -1,5 +1,6 @@
 // mods ~ cross-platforms modules (core/bundler file)
 
 pub mod coreopts;
+pub mod os;
 pub mod panic;
 pub mod ranges;
