+++ tests/by-util/test_logname.rs
@@ -1,5 +1,6 @@
 use crate::common::util::*;
 use std::env;
+use uucore;
 
 #[test]
 fn test_normal() {
@@ -13,7 +14,7 @@ fn test_normal() {
     for (key, value) in env::vars() {
         println!("{}: {}", key, value);
     }
-    if (is_ci() || is_wsl()) && result.stderr.contains("error: no login name") {
+    if (is_ci() || uucore::os::is_wsl_1()) && result.stderr.contains("error: no login name") {
         // ToDO: investigate WSL failure
         // In the CI, some server are failing to return logname.
         // As seems to be a configuration issue, ignoring it
